requirements.txt
pytest==6.1.2
pytest-django==4.1.0
pytest-asyncio==0.14.0
certifi==2018.8.24
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.0.0
django-admin-sortable2==0.7.6
djangorestframework==3.11.0
drfpasswordless==1.3.7
django-ckeditor==6.0.0
future==0.16.0
idna==2.7
Markdown==2.6.11
Pillow==5.2.0
PySocks==1.6.8
python-telegram-bot==10.1.0
psycopg2_binary
pytz==2018.5
requests==2.19.1
urllib3==1.23
gunicorn
uvicorn==0.11.6
discord.py==1.3.4
jsonfield
django-prettyjson==0.4.1
easy-thumbnails==2.6
django-object-actions==1.1.0

#FSM
transitions==0.8.6
django-transitions==0.2

django-redis==4.12.1

# di
dependency-injector==4.5.4

# typing
mypy-extensions==0.4.3
typing-extensions==3.7.4.1
pydantic[email]==1.7.2

# models
pydantic==1.4

freezegun==0.3.12
parameterized==0.7.0

django-simple-history==2.11.0

#
celery==5.0.1
redis==3.2.0

# email:
django-celery-email

# monitoring
sentry-sdk==0.7.2
raven
